fix: return values of decrypt due to different encodings
When an object is first processed, the object is first stringified and encoded with `encodeUTF8` into type `UInt8Array`, which `nacl` uses for all their API.
However, the utf8 encoding is not preserved once `nacl.sign` is called, as the encoding changes to base-64.

Thus, `_decrypt` cannot assume the decrypted object is encoded in UTF8 and thus only returns the raw decrypted data, for processing afterwards.